refactor(search applications): rename engine indices files
