+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/search_application_indices_logic.test.ts
@@ -10,7 +10,7 @@ import { LogicMounter } from '../../../__mocks__/kea_logic';
 import { EnterpriseSearchApplicationDetails } from '../../../../../common/types/search_applications';
 import { FetchSearchApplicationApiLogic } from '../../api/search_applications/fetch_search_application_api_logic';
 
-import { EngineIndicesLogic, EngineIndicesLogicValues } from './engine_indices_logic';
+import { EngineIndicesLogic, EngineIndicesLogicValues } from './search_application_indices_logic';
 
 const DEFAULT_VALUES: EngineIndicesLogicValues = {
   addIndicesFlyoutOpen: false,
